Add first-version React web UI and AG-UI conformance fixes
Adds a standalone React+Vite+TS web chat client in the web/ directory
that integrates with the harnx-serve AG-UI server. The client uses the
stock assistant-ui library with a per-session HttpAgent and relative
URL proxying.

To support the stock @ag-ui/client adapter, this PR also implements
localized protocol conformance fixes in crates/harnx-serve/src/ag_ui.rs.
These changes ensure the server emits RUN_STARTED events before
snapshots, includes synthetic join envelopes, and handles relaxed
string-based session IDs.

Includes a path-filtered GitHub Actions workflow for the web client
and architecture documentation for the AG-UI integration.

Issue: #959

plan: harnx-webui-first-version
