+++ crates/harnx-serve/src/ag_ui.rs
@@ -11,7 +11,7 @@ use crate::session_actor::{
 };
 
 #[cfg(test)]
-use ag_ui_core::{event::RunStartedEvent, types::ids::RunId};
+use ag_ui_core::event::RunStartedEvent;
 use ag_ui_core::{
     event::{
         BaseEvent, CustomEvent, Event, MessagesSnapshotEvent, RunErrorEvent, StepFinishedEvent,
@@ -21,9 +21,11 @@ use ag_ui_core::{
         ToolCallStartEvent,
     },
     types::{
-        ids::{MessageId, ThreadId, ToolCallId},
+        context::Context,
+        ids::{MessageId, RunId, ThreadId, ToolCallId},
         input::RunAgentInput,
         message::{Message as AgUiMessage, Role},
+        tool::Tool,
     },
     JsonValue,
 };
@@ -44,8 +46,93 @@ use http::{Response, StatusCode};
 use http_body_util::{combinators::BoxBody, BodyExt, StreamBody};
 use hyper::body::Frame;
 use tokio::sync::{broadcast, mpsc::UnboundedSender};
-use tokio_stream::{wrappers::BroadcastStream, StreamExt};
+use tokio_stream::wrappers::BroadcastStream;
 use uuid::Uuid;
+
+#[derive(serde::Deserialize)]
+struct RelaxedRunAgentInput<TState = JsonValue> {
+    #[serde(rename = "threadId", default)]
+    thread_id: Option<String>,
+    #[serde(rename = "runId", default)]
+    run_id: Option<String>,
+    messages: Vec<RelaxedAgUiMessage>,
+    #[serde(default)]
+    state: TState,
+    #[serde(default)]
+    tools: Vec<Tool>,
+    #[serde(default)]
+    context: Vec<Context>,
+    #[serde(rename = "forwardedProps", default)]
+    forwarded_props: serde_json::Map<String, serde_json::Value>,
+}
+
+#[derive(serde::Deserialize)]
+#[serde(tag = "role", rename_all = "camelCase")]
+enum RelaxedAgUiMessage {
+    User {
+        #[serde(default)]
+        id: Option<String>,
+        content: String,
+    },
+    Assistant {
+        content: Option<String>,
+        name: Option<String>,
+    },
+    System {
+        content: String,
+    },
+    Developer {
+        content: String,
+    },
+    Tool {
+        #[serde(rename = "toolCallId", default)]
+        tool_call_id: Option<String>,
+        content: String,
+    },
+}
+
+impl From<RelaxedAgUiMessage> for AgUiMessage {
+    fn from(value: RelaxedAgUiMessage) -> Self {
+        match value {
+            RelaxedAgUiMessage::User { id: _, content } => AgUiMessage::User {
+                id: MessageId::random(),
+                content,
+                name: None,
+            },
+            RelaxedAgUiMessage::Assistant { content, name } => AgUiMessage::Assistant {
+                id: MessageId::random(),
+                content,
+                name,
+                tool_calls: None,
+            },
+            RelaxedAgUiMessage::System { content } => AgUiMessage::System {
+                id: MessageId::random(),
+                content,
+                name: None,
+            },
+            RelaxedAgUiMessage::Developer { content } => AgUiMessage::Developer {
+                id: MessageId::random(),
+                content,
+                name: None,
+            },
+            RelaxedAgUiMessage::Tool {
+                tool_call_id,
+                content,
+            } => AgUiMessage::Tool {
+                id: MessageId::random(),
+                content,
+                tool_call_id: tool_call_id
+                    .as_deref()
+                    .map(|id| {
+                        serde_json::from_value(json!(id))
+                            .expect("tool call id should deserialize from string")
+                    })
+                    .unwrap_or_else(ToolCallId::random),
+                error: None,
+            },
+        }
+    }
+}
 
 const THREAD_ID_NAMESPACE: Uuid = Uuid::from_u128(0x9f1f_5b4f_8080_4c1a_9544_1ce1_4b63_1a2f);
 const SSE_KEEPALIVE_INTERVAL: Duration = Duration::from_secs(15);
@@ -480,40 +567,158 @@ fn keep_alive_frame() -> &'static str {
 }
 
 fn keep_alive_stream(interval: Duration) -> impl tokio_stream::Stream<Item = Bytes> {
-    tokio_stream::wrappers::IntervalStream::new({
-        let mut ticker = tokio::time::interval(interval);
-        ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-        ticker
-    })
-    .skip(1)
-    .map(|_| Bytes::from_static(keep_alive_frame().as_bytes()))
+    tokio_stream::StreamExt::map(
+        tokio_stream::StreamExt::skip(
+            tokio_stream::wrappers::IntervalStream::new({
+                let mut ticker = tokio::time::interval(interval);
+                ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
+                ticker
+            }),
+            1,
+        ),
+        |_| Bytes::from_static(keep_alive_frame().as_bytes()),
+    )
 }
 
 pub fn parse_run_input(body: &[u8]) -> Result<RunAgentInput<JsonValue, JsonValue>, AgUiError> {
-    let mut value: serde_json::Value = serde_json::from_slice(body)
-        .map_err(|err| AgUiError::BadRequest(format!("invalid AG-UI JSON body: {err}")))?;
-
+    let value: serde_json::Value = serde_json::from_slice(body)
+        .map_err(|err| AgUiError::BadRequest(format!("invalid AG-UI request body: {err}")))?;
     let obj = value
+        .as_object()
+        .ok_or_else(|| AgUiError::BadRequest("AG-UI body must be a JSON object".to_string()))?;
+    let messages_value = obj.get("messages").cloned().ok_or_else(|| {
+        AgUiError::BadRequest("AG-UI request must include a messages field".to_string())
+    })?;
+    let messages: Vec<AgUiMessage> =
+        serde_json::from_value::<Vec<RelaxedAgUiMessage>>(messages_value)
+            .map_err(|err| AgUiError::BadRequest(format!("invalid AG-UI messages: {err}")))?
+            .into_iter()
+            .map(AgUiMessage::from)
+            .collect();
+
+    let mut relaxed_value = value.clone();
+    let relaxed_obj = relaxed_value
         .as_object_mut()
         .ok_or_else(|| AgUiError::BadRequest("AG-UI body must be a JSON object".to_string()))?;
+    relaxed_obj.insert("messages".to_string(), serde_json::Value::Array(vec![]));
+    relaxed_obj
+        .entry("state".to_string())
+        .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
+    relaxed_obj
+        .entry("tools".to_string())
+        .or_insert_with(|| serde_json::Value::Array(vec![]));
+    relaxed_obj
+        .entry("context".to_string())
+        .or_insert_with(|| serde_json::Value::Array(vec![]));
+    relaxed_obj
+        .entry("forwardedProps".to_string())
+        .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
 
-    if !obj.contains_key("messages") {
-        return Err(AgUiError::BadRequest(
-            "AG-UI request must include a messages field".to_string(),
-        ));
+    let relaxed: RelaxedRunAgentInput<JsonValue> = serde_json::from_value(relaxed_value)
+        .map_err(|err| AgUiError::BadRequest(format!("invalid AG-UI request body: {err}")))?;
+
+    Ok(RunAgentInput {
+        thread_id: relaxed
+            .thread_id
+            .as_deref()
+            .and_then(|value| Uuid::parse_str(value).ok())
+            .map(ThreadId::from)
+            .unwrap_or_else(ThreadId::random),
+        run_id: relaxed
+            .run_id
+            .as_deref()
+            .and_then(|value| Uuid::parse_str(value).ok())
+            .map(RunId::from)
+            .unwrap_or_else(RunId::random),
+        messages,
+        state: relaxed.state,
+        tools: relaxed.tools,
+        context: relaxed.context,
+        forwarded_props: JsonValue::Object(relaxed.forwarded_props),
+    })
+}
+
+fn frame_run_boundary_event(event_type: &str, thread_id: &str, run_id: &str) -> String {
+    let body = serde_json::json!({
+        "type": event_type,
+        "threadId": thread_id,
+        "runId": run_id,
+    });
+    format!("data: {body}\n\n")
+}
+
+fn frame_run_error_event(thread_id: &str, run_id: &str, message: &str) -> String {
+    let body = serde_json::json!({
+        "type": "RUN_ERROR",
+        "threadId": thread_id,
+        "runId": run_id,
+        "message": message,
+    });
+    format!("data: {body}\n\n")
+}
+
+#[derive(Clone, Copy)]
+enum FirstRunState {
+    AwaitingStarted,
+    Active,
+    Complete,
+    Errored,
+}
+
+fn frame_live_event(
+    event: Event,
+    state: &mut FirstRunState,
+    thread_id: &str,
+    run_id: &str,
+) -> Option<Bytes> {
+    match *state {
+        FirstRunState::AwaitingStarted => match event {
+            Event::RunStarted(_) => {
+                *state = FirstRunState::Active;
+                None
+            }
+            Event::RunFinished(_) => {
+                *state = FirstRunState::Complete;
+                Some(Bytes::from(frame_run_boundary_event(
+                    "RUN_FINISHED",
+                    thread_id,
+                    run_id,
+                )))
+            }
+            Event::RunError(err) => {
+                *state = FirstRunState::Errored;
+                Some(Bytes::from(frame_run_error_event(
+                    thread_id,
+                    run_id,
+                    &err.message,
+                )))
+            }
+            other => frame_event(&other).ok().map(Bytes::from),
+        },
+        FirstRunState::Active => match event {
+            Event::RunStarted(_) => None,
+            Event::RunFinished(_) => {
+                *state = FirstRunState::Complete;
+                Some(Bytes::from(frame_run_boundary_event(
+                    "RUN_FINISHED",
+                    thread_id,
+                    run_id,
+                )))
+            }
+            Event::RunError(err) => {
+                *state = FirstRunState::Errored;
+                Some(Bytes::from(frame_run_error_event(
+                    thread_id,
+                    run_id,
+                    &err.message,
+                )))
+            }
+            other => frame_event(&other).ok().map(Bytes::from),
+        },
+        FirstRunState::Complete | FirstRunState::Errored => {
+            frame_event(&event).ok().map(Bytes::from)
+        }
     }
-
-    obj.entry("state".to_string())
-        .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
-    obj.entry("tools".to_string())
-        .or_insert_with(|| serde_json::Value::Array(vec![]));
-    obj.entry("context".to_string())
-        .or_insert_with(|| serde_json::Value::Array(vec![]));
-    obj.entry("forwardedProps".to_string())
-        .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
-
-    serde_json::from_value(value)
-        .map_err(|err| AgUiError::BadRequest(format!("invalid AG-UI request body: {err}")))
 }
 
 fn snapshot_event(messages: Vec<AgUiMessage>) -> Event {
@@ -580,6 +785,108 @@ impl Drop for UnsubscribeOnDrop {
     }
 }
 
+fn build_prompted_event_stream(
+    run_id: &str,
+    thread_id_text: &str,
+    snapshot_frame: Option<Bytes>,
+    live_stream: impl tokio_stream::Stream<Item = Event> + Send + Sync + 'static,
+) -> std::pin::Pin<Box<dyn tokio_stream::Stream<Item = Bytes> + Send + Sync + 'static>> {
+    let run_id = run_id.to_string();
+    let thread_id_text = thread_id_text.to_string();
+    let run_id_for_closure = run_id.clone();
+    let thread_id_text_for_closure = thread_id_text.clone();
+    let live_stream = tokio_stream::StreamExt::filter_map(live_stream, {
+        let mut state = FirstRunState::AwaitingStarted;
+        move |event| {
+            frame_live_event(
+                event,
+                &mut state,
+                &thread_id_text_for_closure,
+                &run_id_for_closure,
+            )
+        }
+    });
+    let remaining = tokio_stream::StreamExt::chain(tokio_stream::iter(snapshot_frame), live_stream);
+    Box::pin(tokio_stream::StreamExt::chain(
+        tokio_stream::once(Bytes::from(frame_run_boundary_event(
+            "RUN_STARTED",
+            &thread_id_text,
+            &run_id,
+        ))),
+        futures_util::stream::select(remaining, keep_alive_stream(SSE_KEEPALIVE_INTERVAL)),
+    ))
+}
+
+fn build_promptless_event_stream(
+    run_id: &str,
+    thread_id_text: &str,
+    snapshot_frame: Option<Bytes>,
+    live_stream: impl tokio_stream::Stream<Item = Event> + Send + Sync + 'static,
+) -> std::pin::Pin<Box<dyn tokio_stream::Stream<Item = Bytes> + Send + Sync + 'static>> {
+    let synthetic = tokio_stream::StreamExt::chain(
+        tokio_stream::StreamExt::chain(
+            tokio_stream::once(Bytes::from(frame_run_boundary_event(
+                "RUN_STARTED",
+                thread_id_text,
+                run_id,
+            ))),
+            tokio_stream::iter(snapshot_frame),
+        ),
+        tokio_stream::once(Bytes::from(frame_run_boundary_event(
+            "RUN_FINISHED",
+            thread_id_text,
+            run_id,
+        ))),
+    );
+    let passthrough =
+        tokio_stream::StreamExt::filter_map(live_stream, |event| match frame_event(&event) {
+            Ok(frame) => Some(Bytes::from(frame)),
+            Err(err) => {
+                log::warn!("failed to serialize AG-UI passthrough frame: {err}");
+                None
+            }
+        });
+    Box::pin(futures_util::stream::select(
+        tokio_stream::StreamExt::chain(synthetic, passthrough),
+        keep_alive_stream(SSE_KEEPALIVE_INTERVAL),
+    ))
+}
+
+fn build_ag_ui_event_stream(
+    handle: &SessionHandle,
+    run_id: &str,
+    thread_id_text: &str,
+    snapshot: Vec<AgUiMessage>,
+    events: broadcast::Receiver<Event>,
+    has_prompt: Option<&str>,
+) -> std::pin::Pin<Box<dyn tokio_stream::Stream<Item = Bytes> + Send + Sync + 'static>> {
+    let snapshot_frame = match frame_event(&snapshot_event(snapshot)) {
+        Ok(frame) => Some(Bytes::from(frame)),
+        Err(err) => {
+            log::warn!("failed to serialize AG-UI snapshot frame: {err}");
+            None
+        }
+    };
+    let handle_for_lag = handle.clone();
+    let live_stream = tokio_stream::StreamExt::then(BroadcastStream::new(events), move |item| {
+        let handle = handle_for_lag.clone();
+        async move {
+            match item {
+                Ok(event) => Some(event),
+                Err(tokio_stream::wrappers::errors::BroadcastStreamRecvError::Lagged(_)) => {
+                    let info = get_info(&handle).await;
+                    Some(snapshot_event(info.history_snapshot))
+                }
+            }
+        }
+    });
+    let live_stream = tokio_stream::StreamExt::filter_map(live_stream, |event| event);
+    match has_prompt {
+        Some(_) => build_prompted_event_stream(run_id, thread_id_text, snapshot_frame, live_stream),
+        None => build_promptless_event_stream(run_id, thread_id_text, snapshot_frame, live_stream),
+    }
+}
+
 pub async fn ag_ui_run_with_call_fn(
     _base_config: &Config,
     registry: &SessionRegistry,
@@ -588,6 +895,12 @@ pub async fn ag_ui_run_with_call_fn(
     req_body: &[u8],
     _call_fn: Option<AgentCallFn>,
 ) -> Result<AppResponse, AgUiError> {
+    let relaxed_run_input: RelaxedRunAgentInput<JsonValue> = serde_json::from_slice(req_body)
+        .map_err(|err| AgUiError::BadRequest(format!("invalid AG-UI request body: {err}")))?;
+    let run_id = relaxed_run_input
+        .run_id
+        .clone()
+        .unwrap_or_else(|| Uuid::new_v4().to_string());
     let run_input = parse_run_input(req_body)?;
     let key = crate::session_actor::SessionKey {
         agent: agent.to_string(),
@@ -595,38 +908,26 @@ pub async fn ag_ui_run_with_call_fn(
     };
     let handle = registry.get_or_spawn(key);
     let SubscribeResult { snapshot, events } = subscribe(&handle).await;
+    let has_prompt = last_user_prompt(&run_input);
+    let thread_id = derive_thread_id(session);
+    let thread_id_text = thread_id.to_string();
 
-    if let Some(text) = last_user_prompt(&run_input) {
-        let _ = prompt(&handle, &text).await;
+    if let Some(text) = has_prompt.as_deref() {
+        let _ = prompt(&handle, text).await;
     }
 
-    let thread_id = derive_thread_id(session);
-    let snapshot_stream = tokio_stream::once(snapshot_event(snapshot));
-    let handle_for_lag = handle.clone();
-    let live_stream = BroadcastStream::new(events)
-        .then(move |item| {
-            let handle = handle_for_lag.clone();
-            async move {
-                match item {
-                    Ok(event) => Some(event),
-                    Err(tokio_stream::wrappers::errors::BroadcastStreamRecvError::Lagged(_)) => {
-                        let info = get_info(&handle).await;
-                        Some(snapshot_event(info.history_snapshot))
-                    }
-                }
-            }
-        })
-        .filter_map(|event| event);
-
+    let event_stream = build_ag_ui_event_stream(
+        &handle,
+        &run_id,
+        &thread_id_text,
+        snapshot,
+        events,
+        has_prompt.as_deref(),
+    );
     let unsubscribe_guard = UnsubscribeOnDrop {
         handle: handle.clone(),
     };
-    let event_stream = snapshot_stream.chain(live_stream).map(|event| {
-        let frame = frame_event(&event).expect("AG-UI event framing should serialize");
-        Bytes::from(frame)
-    });
-    let keep_alive_stream = keep_alive_stream(SSE_KEEPALIVE_INTERVAL);
-    let stream = futures_util::stream::select(event_stream, keep_alive_stream).map(move |frame| {
+    let stream = tokio_stream::StreamExt::map(event_stream, move |frame| {
         let _guard = &unsubscribe_guard;
         Ok::<_, Infallible>(Frame::data(frame))
     });
@@ -790,1625 +1091,5 @@ fn history_role_for_client(role: &Role) -> MessageRole {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_support::TestConfigSandbox;
-    use anyhow::anyhow;
-    use harnx_core::{
-        api_types::CompletionTokenUsage,
-        event::{AgentEventSink, ModelEvent, SessionEvent, TurnEvent},
-        message::MessageContentToolCalls,
-    };
-    use harnx_runtime::{client::ToolCall, config::Config};
-    use http_body_util::BodyExt;
-    use serde_json::{json, Value};
-
-    #[test]
-    fn ag_ui_sink_emits_text_message_content_for_chunk() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id =
-            MessageId::from(uuid::Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap());
-        let sink = super::AgUiSink::with_snapshot(tx, message_id.clone(), false, None);
-
-        sink.emit(
-            AgentEvent::Model(ModelEvent::MessageChunk {
-                blocks: vec![ContentBlock::Text("hello".to_string())],
-            }),
-            None,
-        );
-
-        match rx.try_recv().expect("text start") {
-            Event::TextMessageStart(event) => {
-                assert_eq!(event.message_id, message_id.clone());
-                assert_eq!(event.role, Role::Assistant);
-            }
-            other => panic!("expected TextMessageStart event, got: {other:?}"),
-        }
-        match rx.try_recv().expect("text content") {
-            Event::TextMessageContent(TextMessageContentEvent {
-                base,
-                message_id: mid,
-                delta,
-            }) => {
-                assert_eq!(base.timestamp, None);
-                assert_eq!(base.raw_event, None);
-                assert_eq!(mid, message_id);
-                assert_eq!(delta, "hello");
-            }
-            other => panic!("expected TextMessageContent event, got: {other:?}"),
-        }
-
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_skips_empty_chunk() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::new(tx, message_id);
-
-        sink.emit(
-            AgentEvent::Model(ModelEvent::MessageChunk { blocks: vec![] }),
-            None,
-        );
-        assert!(rx.try_recv().is_err());
-
-        sink.emit(
-            AgentEvent::Model(ModelEvent::MessageChunk {
-                blocks: vec![ContentBlock::Image {
-                    data: vec![],
-                    mime: "image/png".to_string(),
-                }],
-            }),
-            None,
-        );
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_emits_run_error_for_model_error() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::new(tx, message_id);
-
-        sink.emit(
-            AgentEvent::Model(ModelEvent::Error("boom".to_string())),
-            None,
-        );
-
-        let event = rx.try_recv().expect("should receive event");
-        match event {
-            Event::RunError(RunErrorEvent {
-                base,
-                message,
-                code,
-            }) => {
-                assert_eq!(base.timestamp, None);
-                assert_eq!(base.raw_event, None);
-                assert_eq!(message, "boom");
-                assert!(code.is_none());
-            }
-            _ => panic!("expected RunError event, got: {:?}", event),
-        }
-    }
-
-    #[test]
-    fn ag_ui_sink_emits_run_error_for_notice_error() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::new(tx, message_id);
-
-        sink.emit(
-            AgentEvent::Notice(NoticeEvent::Error("fatal error".to_string())),
-            None,
-        );
-
-        let event = rx.try_recv().expect("should receive event");
-        match event {
-            Event::RunError(RunErrorEvent { message, .. }) => {
-                assert_eq!(message, "fatal error");
-            }
-            _ => panic!("expected RunError event, got: {:?}", event),
-        }
-    }
-
-    #[test]
-    fn ag_ui_sink_emits_final_as_content_when_non_empty() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::with_snapshot(tx, message_id.clone(), false, None);
-
-        let usage = CompletionTokenUsage::new(Some(10), Some(5), Some(0));
-        sink.emit(
-            AgentEvent::Model(ModelEvent::Final {
-                output: "final text".to_string(),
-                usage,
-            }),
-            None,
-        );
-
-        match rx.try_recv().expect("text start") {
-            Event::TextMessageStart(event) => {
-                assert_eq!(event.message_id, message_id.clone());
-                assert_eq!(event.role, Role::Assistant);
-            }
-            other => panic!("expected TextMessageStart event, got: {other:?}"),
-        }
-        match rx.try_recv().expect("text content") {
-            Event::TextMessageContent(TextMessageContentEvent {
-                base,
-                message_id: mid,
-                delta,
-            }) => {
-                assert_eq!(mid, message_id);
-                assert_eq!(delta, "final text");
-                assert_eq!(base.timestamp, None);
-            }
-            other => panic!("expected TextMessageContent event, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn ag_ui_sink_skips_final_when_empty() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::new(tx, message_id);
-
-        let usage = CompletionTokenUsage::new(Some(10), Some(5), Some(0));
-        sink.emit(
-            AgentEvent::Model(ModelEvent::Final {
-                output: String::new(),
-                usage,
-            }),
-            None,
-        );
-
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_maps_thought_then_text_and_closes_thinking_before_text() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::with_snapshot(tx, message_id.clone(), false, None);
-
-        sink.emit(
-            AgentEvent::Model(ModelEvent::ThoughtChunk {
-                blocks: vec![ContentBlock::Text("thinking...".to_string())],
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Model(ModelEvent::MessageChunk {
-                blocks: vec![ContentBlock::Text("answer".to_string())],
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Model(ModelEvent::Final {
-                output: String::new(),
-                usage: CompletionTokenUsage::default(),
-            }),
-            None,
-        );
-
-        assert!(matches!(
-            rx.try_recv().expect("thinking start"),
-            Event::ThinkingStart(_)
-        ));
-        match rx.try_recv().expect("thinking text start") {
-            Event::ThinkingTextMessageStart(event) => assert_eq!(event.base.timestamp, None),
-            other => panic!("expected ThinkingTextMessageStart, got: {other:?}"),
-        }
-        match rx.try_recv().expect("thinking delta") {
-            Event::ThinkingTextMessageContent(event) => assert_eq!(event.delta, "thinking..."),
-            other => panic!("expected ThinkingTextMessageContent, got: {other:?}"),
-        }
-        assert!(matches!(
-            rx.try_recv().expect("thinking end"),
-            Event::ThinkingTextMessageEnd(_)
-        ));
-        assert!(matches!(
-            rx.try_recv().expect("thinking close"),
-            Event::ThinkingEnd(_)
-        ));
-        match rx.try_recv().expect("text start") {
-            Event::TextMessageStart(event) => {
-                assert_eq!(event.message_id, message_id);
-                assert_eq!(event.role, Role::Assistant);
-            }
-            other => panic!("expected TextMessageStart, got: {other:?}"),
-        }
-        match rx.try_recv().expect("text delta") {
-            Event::TextMessageContent(event) => {
-                assert_eq!(event.message_id, message_id);
-                assert_eq!(event.delta, "answer");
-            }
-            other => panic!("expected TextMessageContent, got: {other:?}"),
-        }
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_closes_thinking_before_tool_events() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::with_snapshot(tx, message_id.clone(), false, None);
-
-        sink.emit(
-            AgentEvent::Model(ModelEvent::ThoughtChunk {
-                blocks: vec![ContentBlock::Text("planning".to_string())],
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Tool(ToolEvent::Started {
-                id: "tool-1".to_string(),
-                name: "read_history".to_string(),
-                kind: harnx_core::event::ToolKind::Other,
-                markdown: None,
-                input: json!({"limit": 1}),
-                locations: vec![],
-            }),
-            None,
-        );
-
-        assert!(matches!(
-            rx.try_recv().expect("thinking start"),
-            Event::ThinkingStart(_)
-        ));
-        assert!(matches!(
-            rx.try_recv().expect("thinking text start"),
-            Event::ThinkingTextMessageStart(_)
-        ));
-        match rx.try_recv().expect("thinking delta") {
-            Event::ThinkingTextMessageContent(event) => assert_eq!(event.delta, "planning"),
-            other => panic!("expected ThinkingTextMessageContent, got: {other:?}"),
-        }
-        assert!(matches!(
-            rx.try_recv().expect("thinking text end"),
-            Event::ThinkingTextMessageEnd(_)
-        ));
-        assert!(matches!(
-            rx.try_recv().expect("thinking end"),
-            Event::ThinkingEnd(_)
-        ));
-        match rx.try_recv().expect("tool start") {
-            Event::ToolCallStart(event) => {
-                assert_eq!(event.parent_message_id, Some(message_id));
-                assert_eq!(event.tool_call_name, "read_history");
-            }
-            other => panic!("expected ToolCallStart, got: {other:?}"),
-        }
-        assert!(matches!(
-            rx.try_recv().expect("tool args"),
-            Event::ToolCallArgs(_)
-        ));
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_keeps_thinking_open_across_background_session_event() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::with_snapshot(tx, message_id, false, None);
-
-        sink.emit(
-            AgentEvent::Model(ModelEvent::ThoughtChunk {
-                blocks: vec![ContentBlock::Text("thinking one".to_string())],
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Session(SessionEvent::Saved {
-                path: "/tmp/session.json".into(),
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Model(ModelEvent::ThoughtChunk {
-                blocks: vec![ContentBlock::Text(" thinking two".to_string())],
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Turn(TurnEvent::Ended {
-                outcome: harnx_core::event::TurnOutcome::default(),
-            }),
-            None,
-        );
-
-        assert!(matches!(
-            rx.try_recv().expect("thinking start"),
-            Event::ThinkingStart(_)
-        ));
-        assert!(matches!(
-            rx.try_recv().expect("thinking text start"),
-            Event::ThinkingTextMessageStart(_)
-        ));
-        match rx.try_recv().expect("first thinking delta") {
-            Event::ThinkingTextMessageContent(event) => assert_eq!(event.delta, "thinking one"),
-            other => panic!("expected ThinkingTextMessageContent, got: {other:?}"),
-        }
-        match rx.try_recv().expect("saved custom event") {
-            Event::Custom(event) => {
-                assert_eq!(event.name, "session_saved");
-                assert_eq!(event.value, json!({ "path": "/tmp/session.json" }));
-            }
-            other => panic!("expected Custom, got: {other:?}"),
-        }
-        match rx.try_recv().expect("second thinking delta") {
-            Event::ThinkingTextMessageContent(event) => assert_eq!(event.delta, " thinking two"),
-            other => panic!("expected ThinkingTextMessageContent, got: {other:?}"),
-        }
-        assert!(matches!(
-            rx.try_recv().expect("thinking text end"),
-            Event::ThinkingTextMessageEnd(_)
-        ));
-        assert!(matches!(
-            rx.try_recv().expect("thinking end"),
-            Event::ThinkingEnd(_)
-        ));
-        assert!(matches!(
-            rx.try_recv().expect("step finished"),
-            Event::StepFinished(_)
-        ));
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_maps_tool_started_completed_to_ag_ui_events() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::with_snapshot(tx, message_id.clone(), false, None);
-        sink.emit(
-            AgentEvent::Tool(ToolEvent::Started {
-                id: "history-1".to_string(),
-                name: "read_history".to_string(),
-                kind: harnx_core::event::ToolKind::Other,
-                markdown: None,
-                input: json!({"limit": 5, "entry_type": "message"}),
-                locations: vec![],
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Tool(ToolEvent::Completed {
-                id: "history-1".to_string(),
-                output: json!("history checked"),
-                markdown: None,
-            }),
-            None,
-        );
-
-        match rx.try_recv().expect("tool start") {
-            Event::ToolCallStart(event) => {
-                assert_eq!(
-                    event.tool_call_id,
-                    serde_json::from_value(json!("history-1")).unwrap()
-                );
-                assert_eq!(event.tool_call_name, "read_history");
-                assert_eq!(event.parent_message_id, Some(message_id.clone()));
-            }
-            other => panic!("expected ToolCallStart, got: {other:?}"),
-        }
-
-        match rx.try_recv().expect("tool args") {
-            Event::ToolCallArgs(event) => {
-                assert_eq!(
-                    event.tool_call_id,
-                    serde_json::from_value(json!("history-1")).unwrap()
-                );
-                assert_eq!(
-                    event.delta,
-                    json!({"limit": 5, "entry_type": "message"}).to_string()
-                );
-            }
-            other => panic!("expected ToolCallArgs, got: {other:?}"),
-        }
-
-        match rx.try_recv().expect("tool end") {
-            Event::ToolCallEnd(event) => {
-                assert_eq!(
-                    event.tool_call_id,
-                    serde_json::from_value(json!("history-1")).unwrap()
-                );
-            }
-            other => panic!("expected ToolCallEnd, got: {other:?}"),
-        }
-
-        match rx.try_recv().expect("tool result") {
-            Event::ToolCallResult(event) => {
-                assert_eq!(
-                    event.tool_call_id,
-                    serde_json::from_value(json!("history-1")).unwrap()
-                );
-                assert_eq!(event.content, "history checked");
-                assert_eq!(event.role, Role::Tool);
-            }
-            other => panic!("expected ToolCallResult, got: {other:?}"),
-        }
-
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_maps_tool_failures_and_blocked_to_results() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let message_id = MessageId::from(uuid::Uuid::new_v4());
-        let sink = super::AgUiSink::new(tx, message_id);
-
-        sink.emit(
-            AgentEvent::Tool(ToolEvent::Failed {
-                id: "tool-fail".to_string(),
-                error: "boom".to_string(),
-            }),
-            None,
-        );
-        sink.emit(
-            AgentEvent::Tool(ToolEvent::Blocked {
-                id: "tool-blocked".to_string(),
-                name: "danger".to_string(),
-                input: json!({"path": "/tmp"}),
-                reason: "blocked by policy".to_string(),
-            }),
-            None,
-        );
-
-        for (expected_id, expected_content) in
-            [("tool-fail", "boom"), ("tool-blocked", "blocked by policy")]
-        {
-            match rx.try_recv().expect("tool end") {
-                Event::ToolCallEnd(event) => {
-                    assert_eq!(
-                        event.tool_call_id,
-                        serde_json::from_value(json!(expected_id)).unwrap()
-                    );
-                }
-                other => panic!("expected ToolCallEnd, got: {other:?}"),
-            }
-            match rx.try_recv().expect("tool result") {
-                Event::ToolCallResult(event) => {
-                    assert_eq!(
-                        event.tool_call_id,
-                        serde_json::from_value(json!(expected_id)).unwrap()
-                    );
-                    assert_eq!(event.content, expected_content);
-                    assert_eq!(event.role, Role::Tool);
-                }
-                other => panic!("expected ToolCallResult, got: {other:?}"),
-            }
-        }
-
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn frame_event_uses_data_only_sse_format() {
-        let event: Event = Event::RunStarted(RunStartedEvent {
-            base: BaseEvent {
-                timestamp: None,
-                raw_event: None,
-            },
-            thread_id: ThreadId::from(
-                Uuid::parse_str("aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa").unwrap(),
-            ),
-            run_id: RunId::from(Uuid::parse_str("bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb").unwrap()),
-        });
-
-        let framed = frame_event(&event).expect("frame should serialize");
-        assert!(framed.starts_with("data: {\"type\":\"RUN_STARTED\""));
-        assert!(framed.ends_with("\n\n"));
-        assert!(!framed.contains("event:"));
-    }
-
-    #[test]
-    fn parse_run_input_accepts_valid_body() {
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "state": {},
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "hello"
-            }],
-            "tools": [],
-            "context": [],
-            "forwardedProps": {}
-        });
-
-        let parsed = parse_run_input(&serde_json::to_vec(&body).unwrap()).expect("valid body");
-        assert_eq!(parsed.messages.len(), 1);
-    }
-
-    #[test]
-    fn ag_ui_sink_maps_turn_started_and_ended_to_step_events() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let sink = super::AgUiSink::with_snapshot(tx, MessageId::random(), true, None);
-
-        sink.emit(AgentEvent::Turn(TurnEvent::Started), None);
-        sink.emit(
-            AgentEvent::Turn(TurnEvent::Ended {
-                outcome: harnx_core::event::TurnOutcome::default(),
-            }),
-            None,
-        );
-
-        match rx.try_recv().expect("step started") {
-            Event::StepStarted(event) => assert_eq!(event.step_name, "turn-1"),
-            other => panic!("expected StepStarted, got: {other:?}"),
-        }
-        match rx.try_recv().expect("step finished") {
-            Event::StepFinished(event) => assert_eq!(event.step_name, "turn-1"),
-            other => panic!("expected StepFinished, got: {other:?}"),
-        }
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[test]
-    fn ag_ui_sink_maps_plan_event_to_custom() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let sink = super::AgUiSink::with_snapshot(tx, MessageId::random(), true, None);
-
-        sink.emit(
-            AgentEvent::Plan {
-                entries: vec![harnx_core::event::PlanEntry {
-                    status: "in_progress".to_string(),
-                    content: "check logs".to_string(),
-                }],
-            },
-            None,
-        );
-
-        match rx.try_recv().expect("plan custom") {
-            Event::Custom(event) => {
-                assert_eq!(event.name, "plan");
-                assert_eq!(
-                    event.value,
-                    json!([{"status": "in_progress", "content": "check logs"}])
-                );
-            }
-            other => panic!("expected Custom plan, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn ag_ui_sink_maps_compaction_completed_to_custom_and_snapshot() {
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<Event>();
-        let snapshot = Arc::new(|| {
-            vec![AgUiMessage::User {
-                id: MessageId::random(),
-                content: "compacted".to_string(),
-                name: None,
-            }]
-        });
-        let sink = super::AgUiSink::with_snapshot(tx, MessageId::random(), true, Some(snapshot));
-
-        sink.emit(AgentEvent::Session(SessionEvent::CompactingCompleted), None);
-
-        match rx.try_recv().expect("compaction custom") {
-            Event::Custom(event) => {
-                assert_eq!(event.name, "session_compacting_completed");
-                assert_eq!(event.value, json!({}));
-            }
-            other => panic!("expected compaction custom, got: {other:?}"),
-        }
-        match rx.try_recv().expect("snapshot") {
-            Event::MessagesSnapshot(event) => {
-                assert_eq!(event.messages.len(), 1);
-                assert!(
-                    matches!(&event.messages[0], AgUiMessage::User { content, .. } if content == "compacted")
-                );
-            }
-            other => panic!("expected MessagesSnapshot, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn parse_run_input_accepts_minimal_body_without_optional_fields() {
-        // Per plan guardrail: only `messages` is required; state/tools/context/forwardedProps
-        // are semantically optional and should default if omitted.
-        let thread_id = Uuid::new_v4();
-        let run_id = Uuid::new_v4();
-        let body = json!({
-            "threadId": thread_id,
-            "runId": run_id,
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "hello"
-            }]
-        });
-
-        let parsed = parse_run_input(&serde_json::to_vec(&body).unwrap()).expect("minimal body");
-        assert_eq!(parsed.messages.len(), 1);
-        assert_eq!(parsed.thread_id, ThreadId::from(thread_id));
-        assert_eq!(parsed.run_id, RunId::from(run_id));
-    }
-
-    #[test]
-    fn parse_run_input_rejects_missing_messages_field() {
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "state": {},
-            "tools": [],
-            "context": [],
-            "forwardedProps": {}
-        });
-
-        let err = parse_run_input(&serde_json::to_vec(&body).unwrap())
-            .expect_err("missing messages should fail");
-        assert!(matches!(err, AgUiError::BadRequest(msg) if msg.contains("messages")));
-    }
-
-    #[test]
-    fn parse_run_input_accepts_empty_messages_for_join_resume() {
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "state": {},
-            "messages": [],
-            "tools": [],
-            "context": [],
-            "forwardedProps": {}
-        });
-
-        let parsed = parse_run_input(&serde_json::to_vec(&body).unwrap())
-            .expect("empty messages should be allowed for join/resume");
-        assert!(parsed.messages.is_empty());
-    }
-
-    #[test]
-    fn resolve_agent_finds_existing_agent() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("hephaestus", "You are Hephaestus.");
-        let config = sandbox.config();
-
-        let agent = resolve_agent(&config, "hephaestus").expect("agent should resolve");
-        assert_eq!(agent.name(), "hephaestus");
-    }
-
-    #[test]
-    fn resolve_agent_returns_not_found_for_unknown_agent() {
-        let sandbox = TestConfigSandbox::new();
-        let config = sandbox.config();
-
-        let err = resolve_agent(&config, "missing-agent").expect_err("missing agent should fail");
-        assert_eq!(
-            err,
-            AgUiError::NotFound("agent 'missing-agent' not found".to_string())
-        );
-    }
-
-    #[test]
-    fn derive_thread_id_is_deterministic_and_passes_through_uuid() {
-        let non_uuid = "not-a-uuid-session";
-        let derived_a = derive_thread_id(non_uuid);
-        let derived_b = derive_thread_id(non_uuid);
-        assert_eq!(derived_a, derived_b);
-
-        let session_uuid = Uuid::new_v4();
-        let thread_id = derive_thread_id(&session_uuid.to_string());
-        assert_eq!(thread_id, ThreadId::from(session_uuid));
-    }
-
-    #[test]
-    fn build_local_input_sets_agent_and_session_meta() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("hephaestus", "You are Hephaestus.");
-        let base_config = sandbox.config();
-        let prompt_config = harnx_session::fork_prompt_config(&base_config);
-        {
-            let mut config = prompt_config.write();
-            config
-                .use_agent_by_name("hephaestus")
-                .expect("agent should load");
-            config
-                .use_session(Some("session-a"))
-                .expect("session should load");
-        }
-
-        let input = build_local_input(&prompt_config, "hephaestus", "session-a", "hello world")
-            .expect("input should build");
-        assert!(input.with_session());
-        assert!(input.with_agent());
-
-        let session = prompt_config
-            .read()
-            .session
-            .clone()
-            .expect("session should exist");
-        assert_eq!(session.agent_name.as_deref(), Some("hephaestus"));
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_streams_ordered_events_with_stubbed_call_fn() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("hephaestus", "You are Hephaestus.");
-        let config = sandbox.config();
-
-        let run_id_uuid = Uuid::parse_str("22222222-2222-2222-2222-222222222222").unwrap();
-        let session_id = "run-session-1";
-        let req_body = serde_json::to_vec(&json!({
-            "threadId": Uuid::new_v4(),
-            "runId": run_id_uuid,
-            "state": {},
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "forge response"
-            }],
-            "tools": [],
-            "context": [],
-            "forwardedProps": {}
-        }))
-        .unwrap();
-
-        let call_fn: AgentCallFn = Arc::new(|_input, _config, _abort| {
-            Box::pin(async move {
-                harnx_core::sink::emit_agent_event(AgentEvent::Model(ModelEvent::MessageChunk {
-                    blocks: vec![ContentBlock::Text("chunk-text".to_string())],
-                }));
-                Ok((
-                    "assistant final".to_string(),
-                    None,
-                    Vec::<ToolCall>::new(),
-                    CompletionTokenUsage::default(),
-                ))
-            })
-        });
-
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(call_fn),
-        );
-        let response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "hephaestus",
-            session_id,
-            &req_body,
-            None,
-        )
-        .await
-        .expect("AG-UI run should succeed");
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(
-            response
-                .headers()
-                .get("Content-Type")
-                .and_then(|v| v.to_str().ok()),
-            Some("text/event-stream")
-        );
-
-        let parsed_events = read_sse_events_until(response, |events| {
-            events
-                .iter()
-                .any(|event| event["type"].as_str() == Some("RUN_FINISHED"))
-        })
-        .await;
-
-        let event_types = parsed_events
-            .iter()
-            .map(|event| event["type"].as_str().unwrap().to_string())
-            .collect::<Vec<_>>();
-        assert_eq!(
-            event_types,
-            vec![
-                "MESSAGES_SNAPSHOT",
-                "RUN_STARTED",
-                "TEXT_MESSAGE_START",
-                "TEXT_MESSAGE_CONTENT",
-                "TEXT_MESSAGE_END",
-                "RUN_FINISHED",
-            ]
-        );
-
-        assert_eq!(parsed_events[3]["delta"].as_str(), Some("chunk-text"));
-
-        let start_message_id = parsed_events[2]["messageId"].as_str().unwrap().to_string();
-        assert_eq!(
-            parsed_events[3]["messageId"].as_str(),
-            Some(start_message_id.as_str())
-        );
-        assert_eq!(
-            parsed_events[3]["messageId"].as_str(),
-            Some(start_message_id.as_str())
-        );
-
-        let session_messages = load_session_texts(&config, "hephaestus", session_id);
-        assert!(
-            session_messages.iter().any(|text| text == "forge response"),
-            "user prompt should persist via loop"
-        );
-        assert!(
-            session_messages
-                .iter()
-                .any(|text| text == "assistant final"),
-            "assistant output should persist via loop"
-        );
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_emits_run_error_without_run_finished_when_call_fn_fails() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("plain", "You are plain.");
-        let config = sandbox.config();
-        let session_id = "error-path-session";
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "fail deterministically"
-            }]
-        });
-
-        let call_fn: AgentCallFn = Arc::new(|_input, _config, _abort| {
-            Box::pin(async { Err(anyhow!("stubbed call failure")) })
-        });
-
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(call_fn),
-        );
-        let response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&body).unwrap(),
-            None,
-        )
-        .await
-        .expect("ag ui response");
-        let parsed_events = read_sse_events_until(response, |events| {
-            events
-                .iter()
-                .any(|event| event["type"].as_str() == Some("RUN_ERROR"))
-        })
-        .await;
-        let event_types = parsed_events
-            .iter()
-            .map(|event| event["type"].as_str().unwrap().to_string())
-            .collect::<Vec<_>>();
-        assert_eq!(
-            event_types,
-            vec![
-                "MESSAGES_SNAPSHOT",
-                "RUN_STARTED",
-                "TEXT_MESSAGE_START",
-                "RUN_ERROR",
-            ]
-        );
-        assert_eq!(
-            parsed_events[3]["message"].as_str(),
-            Some("stubbed call failure")
-        );
-        assert!(parsed_events
-            .iter()
-            .all(|event| event["type"].as_str() != Some("RUN_FINISHED")));
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_lists_persisted_session_and_history_for_agent() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("plain", "You are plain.");
-        let config = sandbox.config();
-        let session_id = "enumeration-session";
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "enumerate persisted history"
-            }]
-        });
-
-        let call_fn: AgentCallFn = Arc::new(|_input, _config, _abort| {
-            Box::pin(async {
-                let usage = CompletionTokenUsage::new(Some(1), Some(1), Some(0));
-                Ok(("persisted assistant".into(), None, vec![], usage))
-            })
-        });
-
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(call_fn),
-        );
-        let response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&body).unwrap(),
-            None,
-        )
-        .await
-        .expect("persisted run response");
-        assert_run_finished(
-            read_sse_events_until(response, |events| {
-                events
-                    .iter()
-                    .any(|event| event["type"].as_str() == Some("RUN_FINISHED"))
-            })
-            .await,
-        );
-
-        let scoped = super::super::agent_scoped_config(&config, "plain").expect("scoped config");
-        let sessions = super::super::agent_sessions_json(&config, "plain").expect("session list");
-        assert_eq!(sessions.len(), 1);
-        assert!(sessions[0].get("session_id").is_some());
-
-        let loaded = harnx_runtime::config::session::load(
-            &scoped,
-            session_id,
-            &scoped.session_file(session_id),
-        )
-        .expect("load persisted session");
-        assert_eq!(loaded.agent_name.as_deref(), Some("plain"));
-
-        let mut history_iter = loaded.messages.iter();
-        let persisted_user = history_iter
-            .find(|message| message.role.is_user())
-            .expect("persisted user message");
-        let persisted_assistant = history_iter
-            .find(|message| message.role.is_assistant())
-            .expect("persisted assistant message");
-        assert_eq!(
-            persisted_user.content.to_text(),
-            "enumerate persisted history"
-        );
-        assert_eq!(persisted_assistant.content.to_text(), "persisted assistant");
-
-        let shaped = loaded
-            .messages
-            .iter()
-            .filter(|message| message.role.is_user() || message.role.is_assistant())
-            .map(|message| {
-                json!({
-                    "id": format!(
-                        "seq:{}:0",
-                        message.log_seq.expect("persisted history should have log seq")
-                    ),
-                    "role": super::super::history_role_name(message.role),
-                    "content": super::super::history_message_content(message),
-                })
-            })
-            .collect::<Vec<_>>();
-        assert_eq!(
-            shaped,
-            vec![
-                json!({
-                    "id": persisted_user
-                        .log_seq
-                        .map(|seq| format!("seq:{seq}:0"))
-                        .expect("persisted user should have log seq"),
-                    "role": "user",
-                    "content": "enumerate persisted history"
-                }),
-                json!({
-                    "id": persisted_assistant
-                        .log_seq
-                        .map(|seq| format!("seq:{seq}:0"))
-                        .expect("persisted assistant should have log seq"),
-                    "role": "assistant",
-                    "content": "persisted assistant"
-                }),
-            ]
-        );
-
-        let all_agents = Config::all_agents();
-        assert!(all_agents.iter().any(|agent| agent.name() == "plain"));
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_uses_same_message_id_for_wire_and_persistence() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("plain", "You are plain.");
-        let config = sandbox.config();
-        let session_id = "wire-id-session";
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "user1"
-            }]
-        });
-
-        let call_fn: AgentCallFn = Arc::new(|_input, _config, _abort| {
-            Box::pin(async {
-                let usage = CompletionTokenUsage::new(Some(1), Some(1), Some(0));
-                Ok(("assistant1".into(), None, vec![], usage))
-            })
-        });
-
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(call_fn),
-        );
-        let response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&body).unwrap(),
-            None,
-        )
-        .await
-        .expect("run response");
-        let events = read_sse_events_until(response, |events| {
-            events
-                .iter()
-                .any(|event| event["type"].as_str() == Some("RUN_FINISHED"))
-        })
-        .await;
-        assert_run_finished(events.clone());
-
-        let wire_message_ids = events
-            .iter()
-            .filter_map(|event| match event["type"].as_str() {
-                Some("TEXT_MESSAGE_START")
-                | Some("TEXT_MESSAGE_CONTENT")
-                | Some("TEXT_MESSAGE_END") => event["messageId"].as_str().map(str::to_string),
-                _ => None,
-            })
-            .collect::<Vec<_>>();
-        assert!(
-            !wire_message_ids.is_empty(),
-            "expected wire message ids in SSE: {events:?}"
-        );
-        assert!(
-            wire_message_ids.windows(2).all(|ids| ids[0] == ids[1]),
-            "wire message ids should stay stable across SSE events: {wire_message_ids:?}"
-        );
-        let persisted_messages = load_session_messages(&config, "plain", session_id);
-        let persisted_assistant = persisted_messages
-            .iter()
-            .find(|msg| msg.role.is_assistant())
-            .expect("persisted assistant message");
-        assert!(persisted_assistant.id.is_some());
-    }
-
-    fn assert_bad_request_contains(err: &AgUiError, expected: &str) {
-        assert_eq!(err.status_code(), StatusCode::BAD_REQUEST);
-        match err {
-            AgUiError::BadRequest(message) => {
-                assert!(
-                    message.contains(expected),
-                    "expected bad request to contain {expected:?}, got {message:?}"
-                );
-            }
-            other => panic!("expected bad request error, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_idle_join_emits_keep_alive_comment() {
-        tokio::time::pause();
-
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("plain", "You are plain.");
-        let config = sandbox.config();
-        let session_id = "idle-keepalive-session";
-        let registry = crate::session_actor::SessionRegistry::new(config.clone());
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": []
-        });
-
-        let response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&body).unwrap(),
-            None,
-        )
-        .await
-        .expect("idle join response");
-
-        let read_task = tokio::spawn(async move {
-            read_sse_until(
-                response,
-                SSE_KEEPALIVE_INTERVAL + Duration::from_secs(5),
-                |read| !read.events.is_empty() && !read.comments.is_empty(),
-            )
-            .await
-        });
-
-        let mut frame_stream = tokio_stream::once(snapshot_event(Vec::new()))
-            .map(|event| {
-                let frame = frame_event(&event).expect("snapshot frame");
-                Ok::<_, Infallible>(Bytes::from(frame))
-            })
-            .chain(keep_alive_stream(TEST_SSE_KEEPALIVE_INTERVAL).map(Ok::<_, Infallible>));
-
-        let snapshot = tokio::time::timeout(Duration::from_secs(1), frame_stream.next())
-            .await
-            .expect("snapshot timeout")
-            .expect("snapshot frame")
-            .expect("snapshot bytes");
-        let keep_alive = tokio::time::timeout(Duration::from_secs(1), frame_stream.next())
-            .await
-            .expect("keep-alive timeout")
-            .expect("keep-alive frame")
-            .expect("keep-alive bytes");
-
-        let snapshot_text = std::str::from_utf8(&snapshot).expect("snapshot utf8");
-        let keep_alive_text = std::str::from_utf8(&keep_alive).expect("keep-alive utf8");
-        assert!(snapshot_text.starts_with("data: "));
-        assert_eq!(keep_alive_text, keep_alive_frame());
-
-        tokio::task::yield_now().await;
-        tokio::time::advance(SSE_KEEPALIVE_INTERVAL + Duration::from_secs(1)).await;
-        tokio::task::yield_now().await;
-
-        let read = read_task.await.expect("join read task");
-        assert_eq!(read.events[0]["type"], "MESSAGES_SNAPSHOT");
-        assert!(read.comments.iter().any(|frame| frame == ": keep-alive"));
-        assert!(
-            !read
-                .events
-                .iter()
-                .any(|event| event["type"] == ": keep-alive"),
-            "keep-alive comment should not parse as AG-UI event: {:?}",
-            read.events
-        );
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_empty_messages_join_only_snapshot_no_new_run() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("plain", "You are plain.");
-        let config = sandbox.config();
-        let session_id = "join-only-session";
-
-        let first_body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "user1"
-            }]
-        });
-        let first_call_fn: AgentCallFn = Arc::new(|_input, _config, _abort| {
-            Box::pin(async {
-                let usage = CompletionTokenUsage::new(Some(1), Some(1), Some(0));
-                Ok(("assistant1".into(), None, vec![], usage))
-            })
-        });
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(first_call_fn),
-        );
-        let first_response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&first_body).unwrap(),
-            None,
-        )
-        .await
-        .expect("first run");
-        let _ = read_sse_events_until(first_response, |events| {
-            events
-                .iter()
-                .any(|event| event["type"].as_str() == Some("RUN_FINISHED"))
-        })
-        .await;
-
-        let join_body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": []
-        });
-        let response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&join_body).unwrap(),
-            None,
-        )
-        .await
-        .expect("join only");
-        let events = read_sse_events_until(response, |events| !events.is_empty()).await;
-        assert_eq!(events[0]["type"], "MESSAGES_SNAPSHOT");
-        assert!(!events.iter().any(|event| event["type"] == "RUN_STARTED"));
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_uses_only_last_message_user_prompt() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("plain", "You are plain.");
-        let config = sandbox.config();
-        let session_id = "last-message-session";
-
-        let body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": [
-                {
-                    "id": Uuid::new_v4(),
-                    "role": "user",
-                    "content": "ignored earlier"
-                },
-                {
-                    "id": Uuid::new_v4(),
-                    "role": "assistant",
-                    "content": "assistant history"
-                },
-                {
-                    "id": Uuid::new_v4(),
-                    "role": "user",
-                    "content": "last user wins"
-                }
-            ]
-        });
-        let call_fn: AgentCallFn = Arc::new(|input, _config, _abort| {
-            let text = input.text();
-            Box::pin(async move {
-                assert_eq!(text, "last user wins");
-                let usage = CompletionTokenUsage::new(Some(1), Some(1), Some(0));
-                Ok(("assistant2".into(), None, vec![], usage))
-            })
-        });
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(call_fn),
-        );
-        let response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&body).unwrap(),
-            None,
-        )
-        .await
-        .expect("last message prompt run");
-        let events = read_sse_events_until(response, |events| {
-            events
-                .iter()
-                .any(|event| event["type"].as_str() == Some("RUN_FINISHED"))
-        })
-        .await;
-        assert_eq!(events[0]["type"], "MESSAGES_SNAPSHOT");
-        assert!(events.iter().any(|event| event["type"] == "RUN_STARTED"));
-    }
-
-    #[tokio::test]
-    async fn ag_ui_run_resume_same_session_persists_only_new_turn() {
-        let sandbox = TestConfigSandbox::new();
-        sandbox.write_agent("plain", "You are plain.");
-        let config = sandbox.config();
-        let session_id = "resume-session";
-
-        let first_body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": [{
-                "id": Uuid::new_v4(),
-                "role": "user",
-                "content": "user1"
-            }]
-        });
-        let first_call_fn: AgentCallFn = Arc::new(|_input, _config, _abort| {
-            Box::pin(async {
-                let usage = CompletionTokenUsage::new(Some(1), Some(1), Some(0));
-                Ok(("assistant1".into(), None, vec![], usage))
-            })
-        });
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(first_call_fn),
-        );
-        let first_response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&first_body).unwrap(),
-            None,
-        )
-        .await
-        .expect("first run");
-        assert_run_finished(
-            read_sse_events_until(first_response, |events| {
-                events
-                    .iter()
-                    .any(|event| event["type"].as_str() == Some("RUN_FINISHED"))
-            })
-            .await,
-        );
-        let first_messages = load_session_texts(&config, "plain", session_id);
-        assert_eq!(
-            first_messages,
-            vec!["user1".to_string(), "assistant1".to_string()]
-        );
-
-        let persisted_messages = load_session_messages(&config, "plain", session_id);
-        let persisted_assistant_id = persisted_messages
-            .iter()
-            .find(|msg| msg.role.is_assistant())
-            .and_then(|msg| msg.id.clone())
-            .expect("persisted assistant id");
-
-        let second_body = json!({
-            "threadId": Uuid::new_v4(),
-            "runId": Uuid::new_v4(),
-            "messages": [
-                {
-                    "id": MessageId::random(),
-                    "role": "user",
-                    "content": "user1"
-                },
-                {
-                    "id": persisted_assistant_id,
-                    "role": "assistant",
-                    "content": "assistant1"
-                },
-                {
-                    "id": MessageId::random(),
-                    "role": "user",
-                    "content": "user2"
-                }
-            ]
-        });
-        let second_call_fn: AgentCallFn = Arc::new(|_input, _config, _abort| {
-            Box::pin(async {
-                let usage = CompletionTokenUsage::new(Some(1), Some(1), Some(0));
-                Ok(("assistant2".into(), None, vec![], usage))
-            })
-        });
-        let registry = crate::session_actor::SessionRegistry::new_for_tests(
-            config.clone(),
-            std::time::Duration::from_secs(30),
-            Some(second_call_fn),
-        );
-        let second_response = ag_ui_run_with_call_fn(
-            &config,
-            &registry,
-            "plain",
-            session_id,
-            &serde_json::to_vec(&second_body).unwrap(),
-            None,
-        )
-        .await
-        .expect("second run");
-        assert_run_finished(
-            read_sse_events_until(second_response, |events| {
-                events
-                    .iter()
-                    .any(|event| event["type"].as_str() == Some("RUN_FINISHED"))
-            })
-            .await,
-        );
-        let second_messages = load_session_texts(&config, "plain", session_id);
-        assert_eq!(second_messages.len(), first_messages.len() + 2);
-        assert_eq!(
-            second_messages,
-            vec![
-                "user1".to_string(),
-                "assistant1".to_string(),
-                "user2".to_string(),
-                "assistant2".to_string(),
-            ]
-        );
-    }
-
-    fn parse_sse_frames(body: &str) -> Vec<String> {
-        body.split("\n\n")
-            .filter_map(|frame| {
-                let trimmed = frame.trim();
-                if trimmed.is_empty() {
-                    None
-                } else {
-                    Some(trimmed.to_string())
-                }
-            })
-            .collect()
-    }
-
-    async fn parse_sse_events(response: AppResponse) -> Vec<Value> {
-        let body = response
-            .into_body()
-            .collect()
-            .await
-            .expect("collect sse body")
-            .to_bytes();
-        let body_text = std::str::from_utf8(&body).expect("sse utf8");
-        parse_sse_frames(body_text)
-            .into_iter()
-            .map(|frame| parse_sse_frame(&frame))
-            .collect()
-    }
-
-    async fn read_sse_events_until<F>(response: AppResponse, done: F) -> Vec<Value>
-    where
-        F: Fn(&[Value]) -> bool,
-    {
-        read_sse_until(response, std::time::Duration::from_secs(5), |read| {
-            done(&read.events)
-        })
-        .await
-        .events
-    }
-
-    async fn read_sse_until<F>(response: AppResponse, timeout: Duration, done: F) -> SseRead
-    where
-        F: Fn(&SseRead) -> bool,
-    {
-        let mut body = response.into_body().into_data_stream();
-        let mut read = SseRead::default();
-        let mut partial = String::new();
-
-        while !done(&read) {
-            let next = tokio::time::timeout(timeout, body.next())
-                .await
-                .expect("timed out waiting for SSE chunk");
-            let chunk = next
-                .expect("sse stream ended before expected frame")
-                .expect("stream chunk");
-            partial.push_str(std::str::from_utf8(&chunk).expect("sse utf8"));
-
-            while let Some(idx) = partial.find("\n\n") {
-                let frame = partial[..idx].trim().to_string();
-                partial.drain(..idx + 2);
-                if frame.is_empty() {
-                    continue;
-                }
-                read.frames.push(frame.clone());
-                if frame.starts_with(':') {
-                    read.comments.push(frame);
-                } else {
-                    read.events.push(parse_sse_frame(&frame));
-                }
-                if done(&read) {
-                    return read;
-                }
-            }
-        }
-
-        read
-    }
-
-    #[derive(Debug, Default)]
-    struct SseRead {
-        frames: Vec<String>,
-        events: Vec<Value>,
-        comments: Vec<String>,
-    }
-
-    fn parse_sse_frame(frame: &str) -> Value {
-        let payload = frame
-            .strip_prefix("data: ")
-            .expect("sse frame should start with data prefix");
-        serde_json::from_str(payload).expect("frame should be valid json")
-    }
-
-    fn assert_run_finished(events: Vec<Value>) {
-        assert!(
-            events
-                .iter()
-                .any(|event| event["type"].as_str() == Some("RUN_FINISHED")),
-            "expected RUN_FINISHED barrier, got events: {events:?}"
-        );
-    }
-
-    fn load_session_texts(config: &Config, agent: &str, session_id: &str) -> Vec<String> {
-        let prompt_config = harnx_session::fork_prompt_config(config);
-        {
-            let mut cfg = prompt_config.write();
-            cfg.use_agent_by_name(agent).expect("set agent");
-            cfg.use_session(Some(session_id)).expect("load session");
-        }
-        let session_messages = prompt_config
-            .read()
-            .session
-            .as_ref()
-            .expect("session should exist")
-            .messages
-            .iter()
-            .filter(|msg| msg.role.is_user() || msg.role.is_assistant())
-            .map(|msg| msg.content.to_text())
-            .collect();
-        session_messages
-    }
-    fn load_session_messages(config: &Config, agent: &str, session_id: &str) -> Vec<HistoryMsg> {
-        let prompt_config = harnx_session::fork_prompt_config(config);
-        {
-            let mut cfg = prompt_config.write();
-            cfg.use_agent_by_name(agent).expect("set agent");
-            cfg.use_session(Some(session_id)).expect("load session");
-        }
-        let messages = prompt_config
-            .read()
-            .session
-            .as_ref()
-            .expect("session should exist")
-            .messages
-            .iter()
-            .filter(|msg| msg.role.is_user() || msg.role.is_assistant())
-            .cloned()
-            .collect();
-        messages
-    }
-
-    #[test]
-    fn history_messages_for_snapshot_keeps_tool_turn_prose_and_tool_entries() {
-        let tool_call = harnx_core::tool::ToolCall::new(
-            "history".into(),
-            json!({"limit": 5}),
-            Some("call-1".into()),
-            None,
-        );
-        let history = vec![
-            HistoryMsg {
-                role: MessageRole::User,
-                content: MessageContent::Text("prompt".to_string()),
-                id: Some(Uuid::new_v4().to_string()),
-                log_seq: None,
-                log_timestamp: None,
-            },
-            HistoryMsg {
-                role: MessageRole::Assistant,
-                content: MessageContent::ToolCalls(MessageContentToolCalls::new(
-                    vec![harnx_core::tool::ToolResult::new(
-                        tool_call,
-                        json!("tool output"),
-                    )],
-                    "assistant prose".to_string(),
-                    None,
-                )),
-                id: Some(Uuid::new_v4().to_string()),
-                log_seq: None,
-                log_timestamp: None,
-            },
-        ];
-
-        let snapshot = history_messages_for_snapshot(&history);
-        assert_eq!(snapshot.len(), 3);
-        assert!(matches!(&snapshot[0], AgUiMessage::User { content, .. } if content == "prompt"));
-        assert!(
-            matches!(&snapshot[1], AgUiMessage::Assistant { content, .. } if content.as_deref() == Some("assistant prose"))
-        );
-        assert!(
-            matches!(&snapshot[2], AgUiMessage::Tool { content, .. } if content.contains("tool output"))
-        );
-    }
-
-    fn user_msg(content: &str) -> AgUiMessage {
-        AgUiMessage::User {
-            id: MessageId::random(),
-            content: content.to_string(),
-            name: None,
-        }
-    }
-
-    fn assistant_msg(content: &str) -> AgUiMessage {
-        AgUiMessage::Assistant {
-            id: MessageId::random(),
-            content: Some(content.to_string()),
-            name: None,
-            tool_calls: None,
-        }
-    }
-}
+#[path = "ag_ui_tests.rs"]
+mod tests;
