+++ web/.nvmrc
@@ -1,0 +1,1 @@
+v24.12.0
